some csv changes and property handling

+++ src/main/kotlin/no/itverket/batchdemo/dal/PersonRespository.kt
@@ -1,13 +1,14 @@
 package no.itverket.batchdemo.dal
 
 import no.itverket.batchdemo.dal.table.PersonTable
-import org.jetbrains.exposed.sql.*
-import org.jetbrains.exposed.sql.statements.InsertStatement
+import org.jetbrains.exposed.sql.Database
+import org.jetbrains.exposed.sql.StdOutSqlLogger
+import org.jetbrains.exposed.sql.addLogger
+import org.jetbrains.exposed.sql.insert
 import org.jetbrains.exposed.sql.transactions.transaction
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Repository
-import org.springframework.transaction.annotation.Transactional
 import java.time.Instant
 import javax.sql.DataSource
 
@@ -37,17 +38,16 @@ class PersonRespository(dataSource: DataSource) {
             PersonTable.insert {
                 it[firstName] = person.firstName
                 it[lastName] = person.lastName
-                it[age] = person.age
+                it[age] = person.age.toInt()
                 it[email] = person.email
                 it[company] = person.company
                 it[street] = person.street
                 it[city] = person.city
-                it[zipCode] = person.zipCode
+                it[zipCode] = person.zipCode.toShort()
                 it[phone] = person.phone
                 it[creationDate] = Instant.now()
                 it[lastUpdated] = Instant.now()
             }
         }
-
     }
 }
